+++ doc/conf.py
@@ -29,12 +29,12 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "text", "bnf",
                     "demos", "presentations", "scripts", "_generated/bnf/base.rst",
                     "makeHeader"]
 reSubs = [
-    ["λ", r":math:`{\\lambda}`"],
-    ["Δ", r":math:`{\\Delta}`"],
-    ["½", r":math:`{1/2}`"],
-    ["∬", r":math:`{\\iint}`"],
+    ["λ", r":math:`{\\lambda}`", ["MAN_PAGE"]],
+    ["Δ", r":math:`{\\Delta}`", ["MAN_PAGE"]],
+    ["½", r":math:`{1/2}`", ["MAN_PAGE"]],
+    ["∬", r":math:`{\\iint}`", ["MAN_PAGE"]],
     [r".. literalinclude:: ../../doc/bnf/(.*).bnf",
-     r".. include:: ../../doc/_generated/bnf/\1.rst"]
+     r".. include:: ../../doc/_generated/bnf/\1.rst", []]
 ]
 
 for definedType in ("ONEHOT_T", "ONEHOT_AR_T", "PRED_T", "PRED_AR_T",
@@ -42,18 +42,48 @@ for definedType in ("ONEHOT_T", "ONEHOT_AR_T", "PRED_T", "PRED_AR_T",
                     "IMPORTANCE_T", "IMPORTANCE_AR_T", "MODEL_ONEHOT_T",
                     "MOTIF_FLOAT_T"):
     reSubs.append(
-        [definedType, f" :py:data:`{definedType}<bpreveal.internal.constants.{definedType}>` "])
+        [definedType, f" :py:data:`{definedType}<bpreveal.internal.constants.{definedType}>` ", []])
 
 
-def fixLambda(app, what, name, obj, options, lines):  # pylint: disable=unused-argument
+def fixLambda(app, what, name, obj, options, lines):  # pylint: disable=unused-argument  # noqa
     """Replace Greek and other unicode letters in docstrings with something that TeX can handle."""
+    defs = []
+    if app.outdir.parts[-1] == "man":
+        defs.append("MAN_PAGE")
+    printing = [True]
     for i, line in enumerate(lines):
-        for rsub, rrepl in reSubs:
-            line = re.sub(rsub, rrepl, line)  # pylint: disable=redefined-loop-name
-        lines[i] = line
+        match line.split():
+            case ("#define", symbol):
+                defs.append(symbol)
+            case ("#undef", symbol):
+                defs = [x for x in defs if x != symbol]
+            case ("#ifdef", symbol):
+                printing.append(symbol in defs)
+                lines[i] = ""
+            case ("#else",):
+                printing[-1] = not printing[-1]
+                lines[i] = ""
+            case ("#endif",):
+                printing = printing[:-1]
+                lines[i] = ""
+            case ("#ifndef", symbol):
+                printing.append(symbol not in defs)
+                lines[i] = ""
+            case _:
+                if printing[-1]:
+                    for rsub, rrepl, skipEnvs in reSubs:
+                        skip = False
+                        for se in skipEnvs:
+                            if se in defs:
+                                skip = True
+                        if not skip:
+                            line = re.sub(rsub, rrepl, line)  # pylint: disable=redefined-loop-name
+                    lines[i] = line
+                else:
+                    lines[i] = ""
 
 
-def setup(app):
+def setup(app) -> None:  # noqa: ANN001
     """Connect the lambda corrector."""
     app.connect("autodoc-process-docstring", fixLambda)
 
@@ -94,6 +124,38 @@ templates_path = []
 html_theme = "sphinx_rtd_theme"
 html_static_path = ["_generated/static"]
 html_css_files = ["custom-styles.css", "libertinus.css"]
+man1 = ["addNoise", "bestMotifsOnly", "checkJson", "interpretFlat",
+        "interpretPisa", "lengthCalc", "lossWeights", "makeLossPlots",
+        "makePisaFigure", "makePredictions", "metrics", "motifAddQuantiles",
+        "motifScan", "motifSeqletCutoffs", "predictToBigwig", "prepareBed",
+        "prepareTrainingData", "revcompTools", "shapToBigwig", "shapToNumpy",
+        "shiftBigwigs", "shiftPisa", "showModel", "showTrainingProgress",
+        "tileGenome", "trainSoloModel", "trainTransformationModel",
+        "trainCombinedModel"]
+man3 = ["addNoiseUtils", "bedUtils", "callbacks", "colors", "constants",
+        "disableTensorflowLogging", "gaOptimize", "generators",
+        "interpretUtils", "jaccard", "layers", "logUtils", "losses", "models",
+        "motifUtils", "plotting", "plotUtils", "predictUtils", "schema",
+        "slurm", "training", "ushuffle", "utils"]
+man7 = ["bnf", "breakingChanges", "changelog", "countsLossReweighting",
+        "license", "modelArchitectures", "philosophy", "pisa", "programs",
+        "releaseChecklist", "setup", "workflow"]
 
+man_make_section_directory = True
+man_pages = []
+for mp in man1 + man3 + man7:
+    idx = None
+    if mp in man1:
+        idx = 1
+    elif mp in man3:
+        idx = 3
+    elif mp in man7:
+        idx = 7
+    man_pages.append((
+        "_generated/" + mp,
+        mp,
+        "",
+        author,
+        idx))
 # pylint: enable=invalid-name
 # Copyright 2022, 2023, 2024 Charles McAnany. This file is part of BPReveal. BPReveal is free software: You can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 2 of the License, or (at your option) any later version. BPReveal is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with BPReveal. If not, see <https://www.gnu.org/licenses/>.  # noqa  # pylint: disable=line-too-long